requirements.txt
absl-py==2.3.1
asgiref==3.10.0
asn1crypto==1.5.1
astunparse==1.6.3
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deepface==0.0.95
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
docopt==0.6.2
endesive==2.19.1
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
hedera-sdk-py==2.50.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
libclang==18.1.1
lxml==6.0.2
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
mtcnn==1.0.0
namex==0.1.0
num2words==0.5.14
numpy==2.2.6
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.0
pycparser==2.23
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyjnius==1.7.0
PyMuPDF==1.26.6
pypdf==6.1.3
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.4
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf_keras==2.20.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.0
